+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Update affiliate route
+
+### Added
+
 - Create affiliate route
 
 ## [0.1.0] - 2021-11-09